airflow/requirements.txt
snowflake-connector-python
bs4
pandas
pyarrow
python-dotenv
snowflake-sqlalchemy
sqlalchemy
